Convert all internal parameters to control ports
All 12 internal parameters (like hardness, overdrive, fine tuning...)
have been replaced with references to control ports and their values.

This caused me to clean up the code a lot (while at the same time
introducing a little bit more noise during the constructor of the Voice
class, I admit), threwing away weird functions and refactoring quite a
lot of code (at least this is how it felt to me, don't know if this is
supported by diff statistics...).

Exposing control ports is important for the GUI, which works upon
control ports to manipulate the state of the plugin.