sample data as raw audio
The original VST plugin had all its sample data stored in an array. This
bloats the code (mdaEPianoData.cpp was ~2.3MB) and makes it impossible
for users to modify the samples.

The plugin now reads the sample data from a raw audio file
`samples.raw` during initialisation. The mdaEPiano class is responsible
for sample initialisation; voices only get a pointer to the data.

The keygroup / sample map code is now in the mdaEPiano class as this
really doesn't need to be duplicated for every voice.