keep params in synth, pass pointer to voices
Input parameters are no longer changed by the plugin. This rewrite makes
it possible to update dependent internal values when single parameters
are manipulated by the host. This is what the function voice::update(id)
is there for.

All parameters are stored in the synth, and the voices no longer
directly read parameters through p(id). Any change to parameters that
requires recalculation of voice-internal values will trigger the update
function in each voice.